requirements.txt
blinker==1.9.0
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
folium==0.18.0
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
osmtogeojson==0.0.2
psycopg2-binary==2.9.10
requests==2.32.3
shapely==2.0.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
xmltodict==0.14.2
xyzservices==2024.9.0
